Fixed some warnings from running analyze

+++ When Life Gives You Lemons/PostDayView.m
@@ -93,8 +93,8 @@
         summaryView.textAlignment = NSTextAlignmentCenter;
         [summaryView setFont:[UIFont fontWithName:fontName size:fontSize]];
         NSInteger cupsSold = dataStore.getCupsSold;
-        NSAssert(cupsSold >= 0, @"Negative number of cups sold (%d)", cupsSold);
-        NSString *profitFromDay = [NSString stringWithFormat:@"You sold %d cups of lemonade and made $%0.2f.", cupsSold, [profit floatValue]];
+        NSAssert(cupsSold >= 0, @"Negative number of cups sold (%d)", (int) cupsSold);
+        NSString *profitFromDay = [NSString stringWithFormat:@"You sold %d cups of lemonade and made $%0.2f.", (int) cupsSold, [profit floatValue]];
         NSNumber *money = dataStore.getMoney;
         NSAssert(money >= 0, @"Negative money (%@)", money);
         NSString *moneyOnHand = [NSString stringWithFormat:@"Total money on hand: $%0.2f", [money floatValue]];
